apps/worker/pyproject.toml
[tool.poetry]
name = "nutriai-worker"
version = "1.0.0"
description = "AI Meal Plan Generation Worker Service"
authors = ["NutriAI Team <team@nutriai.com>"]
readme = "README.md"
packages = [{include = "worker"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.0.0"
openai = "^1.3.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
